requirements.txt
langchain
langchain-community
langchain-openai
chromadb
fastapi
uvicorn
python-dotenv
beautifulsoup4
lxml
tiktoken
requests
